Namespace, add user state
I was getting colisions (I'm not now) between the animation functions
and libraries that were calling into the animation gear. I insulated the
animation functions inside an animation object.

User state is now piped through the animation gear, starting with an
init() function, so that code can return values that will be used in the
next render call.