Over-clear context in case of translation
Clearing the context with clearRect starting at 0,0 fails if the context
has been translated. This change clears 50% farther out upwards,
downwards, left and right. Typically the context is translated so that
the origin is in the middle of the canvas so that rotations cause the
context to spin in place.